backendproject/requirements.txt
asgiref==3.3.1
dataclasses==0.8
Django==3.0.5
django-rest-framework-mongoengine==3.4.1
django-restframework==0.0.1
djangorestframework==3.12.2
djongo==1.3.4
mongoengine==0.23.0
py-bcrypt==0.4
pymongo==3.11.3
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.2.4
